Factor out images
