+++ client/src/components/ProfilePage/profile.css
@@ -10,7 +10,7 @@
 .header-img {
   width: 100%;
   height: 200px;
-  background: url('./campus_temp.jpg');
+  background: url('../../images/campus_temp.jpg');
   background-size: cover;
 }
